Refactor project to newly established file structure
